+++ vlib/compiler/tests/nameof_test.v
@@ -9,7 +9,7 @@ struct FunkyStruct{ }
 
 fn test_nameof_on_various_types_in_generic() {
 	assert simple(42) == "int"
-	assert simple(3.14) == "f32"
+	assert simple(3.14) == "f64"
 	assert simple("FuBar") == "string"
 	assert simple(FunkyStruct{}) == "FunkyStruct"
 	assert simple(test_nameof_on_various_types_in_generic) == "fn ()"
